slimed up GameActivity and Game with refactor of UI elements into GameUI class
