+++ app/src/main/java/csc_cccix/geocracy/game/GameActivity.java
@@ -3,27 +3,17 @@ package csc_cccix.geocracy.game;
 import android.graphics.PixelFormat;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.MotionEvent;
 import android.view.SurfaceHolder;
 import android.view.Window;
 import android.view.WindowManager;
-import android.view.animation.AlphaAnimation;
-import android.widget.LinearLayout;
 import android.widget.Toast;
-
 import com.jakewharton.rxbinding2.view.RxView;
-
 import csc_cccix.R;
 import csc_cccix.geocracy.EventBus;
 import csc_cccix.geocracy.Util;
-import csc_cccix.geocracy.fragments.CurrentPlayerFragment;
 import csc_cccix.geocracy.fragments.LoadingFragment;
 import csc_cccix.geocracy.game.ui_states.GameAction;
 import csc_cccix.geocracy.game.ui_states.GameEvent;
@@ -38,30 +28,11 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
     public static Game game;
     private GameSurfaceView gameSurfaceView;
 
-    FragmentManager fragmentManager;
-
-    private Fragment activeCurrentPlayerFragment = null;
-    private Fragment activeBottomPaneFragment = null;
-    private Fragment activeOverlayFragment = null;
-
-    private CompositeDisposable disposables;
-
-    public FloatingActionButton endTurnButton;
-    public FloatingActionButton attackBtn;
-    public FloatingActionButton addUnitBtn;
-    public FloatingActionButton removeUnitBtn;
-    public FloatingActionButton cancelBtn;
-    public FloatingActionButton gameInfoBtn;
-    public FloatingActionButton settingBtn;
-    public FloatingActionButton closeOverlayBtn;
-    public FloatingActionButton confirmButton;
-    public FloatingActionButton fortifyButton;
-
+    public CompositeDisposable disposables;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        boolean fromGameLoad;
 
         disposables = new CompositeDisposable();
 
@@ -78,10 +49,9 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
         gameSurfaceView = findViewById(R.id.gameplaySurfaceView);
         gameSurfaceView.getHolder().addCallback(this);
 
-        fragmentManager = getSupportFragmentManager();
+        EventBus.subscribe("SAVE_GAME_EVENT", this, event -> handleSaveEvent((GameEvent) event));
 
         // Setup game
-        fromGameLoad = false;
         Boolean load = (Boolean)getIntent().getSerializableExtra("GAME_LOAD");
         // Load game
         if (load != null && load) {
@@ -90,8 +60,6 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
             if (game == null) {
                 Log.e("", "Failed to load game");
                 Util.exit();
-            } else {
-                fromGameLoad = true;
             }
         }
         // Start new game
@@ -103,102 +71,7 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
             game = new Game(this, playerName, numPlayers, Util.colorToVec3(mainPlayerColor), seed);
         }
 
-        // Get Layout Frame +
-        CoordinatorLayout frame = findViewById(R.id.gameLayout);
-        LinearLayout uiLayout = new LinearLayout(this);
-        uiLayout.setOrientation(LinearLayout.VERTICAL);
-
-        cancelBtn = findViewById(R.id.cancelBtn);
-        cancelBtn.hide();
-        disposables.add(RxView.touches(cancelBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.CANCEL_TAPPED, null));
-            }
-        }));
-
-        confirmButton = findViewById(R.id.confirmButton);
-        confirmButton.hide();
-        disposables.add(RxView.touches(confirmButton).subscribe(event -> {
-            if (event.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.CONFIRM_TAPPED, null));
-            }
-        }));
-
-        endTurnButton = findViewById(R.id.endTurnButton);
-        endTurnButton.hide();
-        disposables.add(RxView.touches(endTurnButton).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.END_TURN_TAPPED, null));
-            }
-        }));
-
-        attackBtn = findViewById(R.id.attackBtn);
-        attackBtn.hide();
-        disposables.add(RxView.touches(attackBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.ATTACK_TAPPED, null));
-            }
-        }));
-
-        fortifyButton = findViewById(R.id.fortifyButton);
-        fortifyButton.hide();
-        disposables.add(RxView.touches(fortifyButton).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.FORTIFY_TAPPED, null));
-            }
-        }));
-
-        addUnitBtn = findViewById(R.id.addUnitBtn);
-        addUnitBtn.hide();
-        disposables.add(RxView.touches(addUnitBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.ADD_UNIT_TAPPED, null));
-            }
-        }));
-
-        removeUnitBtn = findViewById(R.id.removeUnitBtn);
-        removeUnitBtn.hide();
-        disposables.add(RxView.touches(removeUnitBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.REMOVE_UNIT_TAPPED, null));
-            }
-        }));
-
-        settingBtn = findViewById(R.id.inGameSettingsBtn);
-        settingBtn.show();
-        disposables.add(RxView.touches(settingBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.SETTINGS_TAPPED, null));
-            }
-        }));
-
-        gameInfoBtn = findViewById(R.id.gameInfoBtn);
-        gameInfoBtn.show();
-        disposables.add(RxView.touches(gameInfoBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.GAME_INFO_TAPPED, null));
-            }
-        }));
-
-        closeOverlayBtn = findViewById(R.id.closeOverlayBtn);
-        closeOverlayBtn.hide();
-
-        disposables.add(RxView.touches(closeOverlayBtn).subscribe(e -> {
-            if (e.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                EventBus.publish(USER_ACTION, new GameEvent(GameAction.CLOSE_OVERLAY_TAPPED, null));
-            }
-        }));
-
-        frame.addView(uiLayout);
-        game.showOverlayFragment(new LoadingFragment());
-
-        EventBus.subscribe("SAVE_GAME_EVENT", this, event -> handleSaveEvent((GameEvent) event));
-
-//        if (fromGameLoad) {
-//            game.setActivityAndState(this, new DefaultState(game));
-//        } else {
-//            game.setActivityAndState(this, new SetUpInitTerritoriesState(game));
-//        }
+//        game.UI.showOverlayFragment(new LoadingFragment());
 
     }
 
@@ -212,119 +85,6 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
                 runOnUiThread(() -> Toasty.info(this, "Error saving game", Toast.LENGTH_LONG).show());
             }
         }
-    }
-
-    public void updateCurrentPlayerFragment() {
-        CurrentPlayerFragment currentPlayerFragment = CurrentPlayerFragment.newInstance(game.getCurrentPlayer());
-        FragmentTransaction ft = fragmentManager.beginTransaction();
-        if (activeCurrentPlayerFragment != null) {
-            ft.remove(activeCurrentPlayerFragment);
-        }
-        if (activeCurrentPlayerFragment != currentPlayerFragment) {
-            ft.add(R.id.gameLayout, currentPlayerFragment);
-            activeCurrentPlayerFragment = currentPlayerFragment;
-        }
-        ft.commit();
-    }
-
-//    public void showOverlayFragment(Fragment overlayFragment) {
-//        removeActiveBottomPaneFragment();
-//
-//        FragmentTransaction ft = fragmentManager.beginTransaction();
-//        ft.add(R.id.gameLayout, overlayFragment);
-//        ft.commit();
-//
-//        activeOverlayFragment = overlayFragment;
-//        this.runOnUiThread(() -> {
-//            if (overlayFragment.getClass() == CurrentPlayerFragment.class) {
-//                closeOverlayBtn.hide();
-//                settingBtn.show();
-//                gameInfoBtn.show();
-//            } else {
-//                closeOverlayBtn.show();
-//                settingBtn.hide();
-//                gameInfoBtn.hide();
-//            }
-//        });
-//    }
-//
-//    public void removeActiveOverlayFragment() {
-//        if (activeOverlayFragment != null) {
-//            FragmentTransaction ft = fragmentManager.beginTransaction();
-//            ft.remove(activeOverlayFragment);
-//            ft.commit();
-//            activeOverlayFragment = null;
-//        }
-//        this.runOnUiThread(() -> {
-//            closeOverlayBtn.hide();
-//            settingBtn.show();
-//            gameInfoBtn.show();
-//        });
-//    }
-
-     public void showBottomPaneFragment(Fragment bottomPaneFragment) {
-        removeActiveBottomPaneFragment();
-        FragmentTransaction ft = fragmentManager.beginTransaction();
-        ft.add(R.id.gameLayout, bottomPaneFragment);
-        ft.commit();
-        activeBottomPaneFragment = bottomPaneFragment;
-    }
-
-     public void removeActiveBottomPaneFragment() {
-        if (activeBottomPaneFragment != null) {
-            FragmentTransaction ft = fragmentManager.beginTransaction();
-            ft.remove(activeBottomPaneFragment);
-            ft.commit();
-            activeBottomPaneFragment = null;
-        }
-    }
-
-    public void setAttackModeButtonVisibilityAndActiveState(boolean isVisible, boolean isActive) {
-        setFABVisibilityAndActiveState(attackBtn, isVisible, isActive);
-    }
-
-    public void setConfirmButtonVisibilityAndActiveState(boolean isVisible, boolean isActive) {
-        setFABVisibilityAndActiveState(confirmButton, isVisible, isActive);
-    }
-
-    public void setFortifyButtonVisibilityAndActiveState(boolean isVisible, boolean isActive) {
-        setFABVisibilityAndActiveState(fortifyButton, isVisible, isActive);
-    }
-
-    private void setFABVisibilityAndActiveState(FloatingActionButton fab, boolean isVisible, boolean isActive) {
-        this.runOnUiThread(() -> {
-            AlphaAnimation alphaChange;
-            if (isActive) { alphaChange = new AlphaAnimation(fab.getAlpha(), 1.0f); }
-            else { alphaChange = new AlphaAnimation(fab.getAlpha(), 0.4f); }
-            alphaChange.setFillAfter(true);
-            fab.startAnimation(alphaChange);
-            if (isVisible) { fab.show(); }
-            else { fab.hide(); }
-        });
-    }
-
-    public void setUpdateUnitCountButtonsVisibility(boolean isVisible) {
-        this.runOnUiThread(() -> {
-            if (isVisible) {
-                addUnitBtn.show();
-                removeUnitBtn.show();
-            } else {
-                addUnitBtn.hide();
-                removeUnitBtn.hide();
-            }
-        });
-    }
-
-    public void hideAllGameInteractionButtons() {
-        this.runOnUiThread(() -> {
-            attackBtn.hide();
-            cancelBtn.hide();
-            addUnitBtn.hide();
-            removeUnitBtn.hide();
-            endTurnButton.hide();
-            confirmButton.hide();
-            fortifyButton.hide();
-        });
     }
 
 
@@ -357,7 +117,7 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
                 EventBus.publish("WORLD_TOUCH_EVENT", e);
             }
         }));
-        new Handler().postDelayed(game::removeOverlayFragment, 4000);
+        new Handler().postDelayed(game.UI::removeOverlayFragment, 4000);
     }
 
     @Override
@@ -370,37 +130,5 @@ public class GameActivity extends AppCompatActivity implements SurfaceHolder.Cal
         // do nothing for now
     }
 
-    public FloatingActionButton getAddUnitBtn() {
-        return addUnitBtn;
-    }
 
-    public FloatingActionButton getAttackBtn() {
-        return attackBtn;
-    }
-
-    public FloatingActionButton getCancelBtn() {
-        return cancelBtn;
-    }
-
-    public FloatingActionButton getConfirmButton() { return confirmButton; }
-
-    public FloatingActionButton getCloseOverlayBtn() {
-        return closeOverlayBtn;
-    }
-
-    public FloatingActionButton getEndTurnButton() {
-        return endTurnButton;
-    }
-
-    public FloatingActionButton getGameInfoBtn() {
-        return gameInfoBtn;
-    }
-
-    public FloatingActionButton getRemoveUnitBtn() {
-        return removeUnitBtn;
-    }
-
-    public FloatingActionButton getSettingBtn() {
-        return settingBtn;
-    }
 }
