+++ app/src/main/java/csc_cccix/geocracy/game/ui_states/FortifyTerritoryState.java
@@ -39,9 +39,9 @@ public class FortifyTerritoryState extends IGameplayState {
         }
 
         SM.Game.getActivity().runOnUiThread(() -> {
-            SM.Game.getActivity().hideAllGameInteractionButtons();
-            SM.Game.getActivity().getConfirmButton().show();
-            SM.Game.getActivity().getCancelBtn().show();
+            SM.Game.UI.hideAllGameInteractionButtons();
+            SM.Game.UI.getConfirmButton().show();
+            SM.Game.UI.getCancelBtn().show();
         });
 
     }
@@ -50,9 +50,9 @@ public class FortifyTerritoryState extends IGameplayState {
     public void DeinitializeState() {
         Log.i(TAG, "DEINIT STATE");
 
-        SM.Game.removeActiveBottomPaneFragment();
+        SM.Game.UI.removeActiveBottomPaneFragment();
         SM.Game.getActivity().runOnUiThread(() -> {
-            SM.Game.getActivity().hideAllGameInteractionButtons();
+            SM.Game.UI.hideAllGameInteractionButtons();
         });
     }
 
@@ -72,8 +72,8 @@ public class FortifyTerritoryState extends IGameplayState {
                         SM.Game.getWorld().selectTerritory(originTerritory);
                         SM.Game.getWorld().targetTerritory(destinationTerritory);
                         SM.Game.getCameraController().targetTerritory(destinationTerritory);
-                        SM.Game.getActivity().runOnUiThread(() -> SM.Game.getActivity().setUpdateUnitCountButtonsVisibility(true));
-                        SM.Game.showBottomPaneFragment(FortifyTerritoryFragment.newInstance(originTerritory, destinationTerritory));
+                        SM.Game.getActivity().runOnUiThread(() -> SM.Game.UI.setUpdateUnitCountButtonsVisibility(true));
+                        SM.Game.UI.showBottomPaneFragment(FortifyTerritoryFragment.newInstance(originTerritory, destinationTerritory));
                     } else {
                         SM.Game.getActivity().runOnUiThread(() -> Toasty.info(SM.Game.getActivity().getBaseContext(), "Cannot move units to another players territory!.", Toast.LENGTH_LONG).show());
                     }
@@ -123,8 +123,8 @@ public class FortifyTerritoryState extends IGameplayState {
         }
 
         SM.Game.getActivity().runOnUiThread(() -> {
-            SM.Game.getActivity().getConfirmButton().show();
-            SM.Game.getActivity().getCancelBtn().hide();
+            SM.Game.UI.getConfirmButton().show();
+            SM.Game.UI.getCancelBtn().hide();
         });
     }
 }
